refactor and document precise data models
