+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "cwbc-odkx-app",
-  "version": "1.0.0-beta.4",
+  "version": "1.0.0-beta.5",
   "description": "",
   "main": "index.js",
   "scripts": {
